3. Интерфейс Input [#181092 #222936]

+++ src/main/java/ru/job4j/poly/Transport.java
@@ -5,5 +5,5 @@ public interface Transport {
 
     void passengers(int amountPassenger);
 
-    int refuel(int amountFuel);
+    double refuel(double amountFuel);
 }
